requirements.txt
pymongo[tls]
python-dateutil
pytz
requests
bson
lambda-decorators